requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
pandas>=2.0.0
numpy>=1.24.0
